당첨 번호 입력을 분리 하고 테스트 추가
